Initial inclusion of SharedViewModel class and some name refactoring
