+++ PhotoPoints/app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             android:value="AIzaSyCyiBxrBCJYCqZwgUBT0O0B7tC_7FqLc68" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme">
         </activity>
