+++ PhotoPoints/app/src/main/java/edu/cascadia/mobas/photopoints/ui/SharedViewModel.java
@@ -1,0 +1,32 @@
+package edu.cascadia.mobas.photopoints.ui;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class SharedViewModel extends ViewModel {
+
+    private int qrScanItemId;
+    private int detailsItemId;
+
+
+    public SharedViewModel() {
+        // viewmodel initialization
+    }
+
+
+    public int getDetailsItemId() {
+        return detailsItemId;
+    }
+
+    public void setDetailsItemId(int id) {
+        detailsItemId = id;
+    }
+
+    public int getQrScalItemId() {
+        return qrScanItemId;
+    }
+
+    public void setQrScanItemId(int id)  {
+        qrScanItemId = id;
+    }
+
+}
